+++ gre/helpers/error.ts
@@ -1,5 +1,5 @@
 import { HardhatPluginError } from 'hardhat/plugins'
-import { logError } from '../logger'
+import { logError } from './logger'
 
 export class GREPluginError extends HardhatPluginError {
   constructor(message: string) {
